+++ src/main/kotlin/com/github/intheclouddan/intellijpluginld/notifications/Notifications.kt
@@ -1,13 +1,15 @@
 package com.github.intheclouddan.intellijpluginld.notifications
 
 import com.intellij.notification.Notification
-import com.intellij.notification.NotificationDisplayType
-import com.intellij.notification.NotificationGroup
+import com.intellij.notification.NotificationGroupManager
 import com.intellij.notification.NotificationType
 import com.intellij.openapi.project.Project
 
 class ConfigNotifier {
-    private val notificationGroup = NotificationGroup("Launchdarkly Configuration", NotificationDisplayType.BALLOON, true)
+    private val notificationGroup = NotificationGroupManager.getInstance()
+        .getNotificationGroup(
+            "LaunchDarkly Configuration"
+        )
 
     fun notify(project: Project?, content: String?): Notification {
         val notification: Notification = notificationGroup.createNotification(content!!, NotificationType.ERROR)